futures_concurrency::array
Struct Merge

pub struct Merge<S, const N: usize>
where
    S: Stream,
{ /* private fields */ }

A stream that merges multiple streams into a single stream.

This struct is created by the merge method on the Merge trait. See its documentation for more.