futures_concurrency::future
Trait RaceOk

pub trait RaceOk {
    type Output;
    type Error;
    type Future: Future<Output = Result<Self::Output, Self::Error>>;

    // Required method
    fn race_ok(self) -> Self::Future;
}

Wait for the first successful future to complete.

Awaits multiple futures simultaneously, returning the output of the first future which completes. If no future completes successfully, returns an aggregate error of all failed futures.

Required Associated Types

type Output

The resulting output type.

type Error

The resulting error type.

type Future: Future<Output = Result<Self::Output, Self::Error>>

Which kind of future are we turning this into?

Required Methods

fn race_ok(self) -> Self::Future

Waits for the first successful future to complete.