futures_concurrency::future
Trait TryJoin

pub trait TryJoin {
    type Output;
    type Error;
    type Future: Future<Output = Result<Self::Output, Self::Error>>;

    // Required method
    fn try_join(self) -> Self::Future;
}

Wait for all futures to complete successfully, or abort early on error.

In the case a future errors, all other futures will be cancelled. If futures have been completed, their results will be discarded.

If you want to keep partial data in the case of failure, see the merge operation.

Required Associated Types

type Output

The resulting output type.

type Error

The resulting error type.

type Future: Future<Output = Result<Self::Output, Self::Error>>

Which kind of future are we turning this into?

Required Methods

fn try_join(self) -> Self::Future

Waits for multiple futures to complete, either returning when all futures complete successfully, or return early when any future completes with an error.