futures_concurrency::vec
Struct Join

pub struct Join<Fut>
where
    Fut: Future,
{ /* private fields */ }

A future which waits for multiple futures to complete.

This struct is created by the join method on the Join trait. See its documentation for more.